postgres-db: split mem, meta, memento
