+++ packages/postgres-db/index.ts
@@ -3,5 +3,7 @@
 export * from './src/cleanUpLastUserMem'
 export * from './src/getConversation'
 export * from './src/mem'
+export * from './src/memento'
+export * from './src/meta'
 export * from './src/postgresdb'
 export * from './src/rawQueries'
